♻️ GM7HX4 guard: split commit orchestration helpers
