+++ .agentplane/tasks/202604231613-GM7HX4/README.md
@@ -1,0 +1,94 @@
+---
+id: "202604231613-GM7HX4"
+title: "Decompose guard commit runtime hotspot"
+status: "DOING"
+priority: "high"
+owner: "CODER"
+revision: 4
+origin:
+  system: "manual"
+depends_on: []
+tags:
+  - "code"
+verify: []
+plan_approval:
+  state: "approved"
+  updated_at: "2026-04-23T16:13:43.889Z"
+  updated_by: "ORCHESTRATOR"
+  note: null
+verification:
+  state: "pending"
+  updated_at: null
+  updated_by: null
+  note: null
+commit: null
+comments:
+  -
+    author: "CODER"
+    body: "Start: decompose guard commit runtime hotspot into local helper modules for staging, cache reset, and follow-up refresh commits while preserving guard commit and close-commit behavior."
+events:
+  -
+    type: "status"
+    at: "2026-04-23T16:13:44.541Z"
+    author: "CODER"
+    from: "TODO"
+    to: "DOING"
+    note: "Start: decompose guard commit runtime hotspot into local helper modules for staging, cache reset, and follow-up refresh commits while preserving guard commit and close-commit behavior."
+doc_version: 3
+doc_updated_at: "2026-04-23T16:13:44.552Z"
+doc_updated_by: "CODER"
+description: "Refactor packages/agentplane/src/commands/guard/impl/commit.ts into smaller local modules without changing guard commit or close-commit behavior. Keep CLI contract stable, add focused regression coverage, and verify hotspot reduction."
+sections:
+  Summary: |-
+    Decompose guard commit runtime hotspot
+    
+    Refactor packages/agentplane/src/commands/guard/impl/commit.ts into smaller local modules without changing guard commit or close-commit behavior. Keep CLI contract stable, add focused regression coverage, and verify hotspot reduction.
+  Scope: |-
+    - In scope: Refactor packages/agentplane/src/commands/guard/impl/commit.ts into smaller local modules without changing guard commit or close-commit behavior. Keep CLI contract stable, add focused regression coverage, and verify hotspot reduction.
+    - Out of scope: unrelated refactors not required for "Decompose guard commit runtime hotspot".
+  Plan: "Split packages/agentplane/src/commands/guard/impl/commit.ts into smaller helper modules for task-index cache reset, staging helpers, and task-artifact refresh commit flow; keep cmdCommit/cmdCloseCommit behavior and CLI output stable; verify with focused guard tests plus typecheck and hotspot check."
+  Verify Steps: |-
+    1. Review the changed artifact or behavior for the `code` task. Expected: the requested outcome is visible and matches the approved scope.
+    2. Run the most relevant validation step for the `code` task. Expected: it succeeds without unexpected regressions in touched scope.
+    3. Compare the final result against the task summary and scope. Expected: any remaining follow-up is explicit in ## Findings.
+  Verification: |-
+    <!-- BEGIN VERIFICATION RESULTS -->
+    <!-- END VERIFICATION RESULTS -->
+  Rollback Plan: |-
+    - Revert task-related commit(s).
+    - Re-run required checks to confirm rollback safety.
+  Findings: ""
+id_source: "generated"
+---
+## Summary
+
+Decompose guard commit runtime hotspot
+
+Refactor packages/agentplane/src/commands/guard/impl/commit.ts into smaller local modules without changing guard commit or close-commit behavior. Keep CLI contract stable, add focused regression coverage, and verify hotspot reduction.
+
+## Scope
+
+- In scope: Refactor packages/agentplane/src/commands/guard/impl/commit.ts into smaller local modules without changing guard commit or close-commit behavior. Keep CLI contract stable, add focused regression coverage, and verify hotspot reduction.
+- Out of scope: unrelated refactors not required for "Decompose guard commit runtime hotspot".
+
+## Plan
+
+Split packages/agentplane/src/commands/guard/impl/commit.ts into smaller helper modules for task-index cache reset, staging helpers, and task-artifact refresh commit flow; keep cmdCommit/cmdCloseCommit behavior and CLI output stable; verify with focused guard tests plus typecheck and hotspot check.
+
+## Verify Steps
+
+1. Review the changed artifact or behavior for the `code` task. Expected: the requested outcome is visible and matches the approved scope.
+2. Run the most relevant validation step for the `code` task. Expected: it succeeds without unexpected regressions in touched scope.
+3. Compare the final result against the task summary and scope. Expected: any remaining follow-up is explicit in ## Findings.
+
+## Verification
+
+<!-- BEGIN VERIFICATION RESULTS -->
+<!-- END VERIFICATION RESULTS -->
+
+## Rollback Plan
+
+- Revert task-related commit(s).
+- Re-run required checks to confirm rollback safety.
+
+## Findings
