+++ packages/agentplane/src/commands/guard/impl/commit.ts
@@ -1,16 +1,8 @@
-import { mkdir, rm, writeFile } from "node:fs/promises";
-import path from "node:path";
-
-import { buildTaskArtifactRefreshCommitSubject } from "@agentplaneorg/core/commit";
-
-import { resolveTaskIndexPath } from "../../../backends/task-index.js";
 import { mapCoreError } from "../../../cli/error-map.js";
 import { infoMessage, successMessage } from "../../../cli/output.js";
 import { withDiagnosticContext } from "../../shared/diagnostics.js";
 import { CliError } from "../../../shared/errors.js";
-import { protectedPathKindForFile } from "../../../shared/protected-paths.js";
 import { refreshBranchPrArtifactsAfterTaskCommit } from "../../shared/post-commit-pr-artifacts.js";
-import { isTaskLocalAdvancePath } from "../../shared/task-local-freshness.js";
 import {
   loadCommandContext,
   loadTaskFromContext,
@@ -24,180 +16,18 @@ import { resolveIgnoredDirectCloseDirtyPaths } from "./close-dirt.js";
 import { buildCloseCommitMessage, taskReadmePathForTask } from "./close-message.js";
 import { asCommitFailure } from "./commit-diagnostics.js";
 import { buildGitCommitEnv, resolveCanonicalGitIdentity } from "./env.js";
+import {
+  commitRefreshedTaskArtifacts,
+  formatCommitRef,
+  refreshBranchPrArtifactsForCloseCommit,
+  resetRebuildableTaskIndexCache,
+} from "./commit-refresh.js";
+import {
+  hasExplicitCommitScope,
+  stageActiveTaskArtifactsFromAllowTasks,
+  stageCloseCommitPaths,
+} from "./commit-stage.js";
 import { guardCommitCheck } from "./policy.js";
-
-async function resetRebuildableTaskIndexCache(ctx: CommandContext): Promise<void> {
-  const gitRoot = path.resolve(ctx.resolvedProject.gitRoot);
-  const workflowDirAbs = path.resolve(gitRoot, ctx.config.paths.workflow_dir);
-  const cachePath = path.resolve(resolveTaskIndexPath(workflowDirAbs));
-  const relativeCachePath = path.relative(gitRoot, cachePath);
-  if (!relativeCachePath || relativeCachePath.startsWith("..")) return;
-
-  try {
-    await execFileAsync("git", ["ls-files", "--error-unmatch", "--", relativeCachePath], {
-      cwd: gitRoot,
-      env: gitEnv(),
-    });
-    const gitPath = relativeCachePath.split(path.sep).join("/");
-    const cacheBlob = await execFileAsync("git", ["show", `HEAD:${gitPath}`], {
-      cwd: gitRoot,
-      env: gitEnv(),
-    });
-    await mkdir(path.dirname(cachePath), { recursive: true });
-    await writeFile(cachePath, cacheBlob.stdout, "utf8");
-    await execFileAsync("git", ["restore", "--staged", "--", relativeCachePath], {
-      cwd: gitRoot,
-      env: gitEnv(),
-    });
-  } catch {
-    await rm(cachePath, { force: true });
-  }
-  ctx.git.invalidateStatus();
-}
-
-async function stageCloseCommitPaths(opts: {
-  ctx: CommandContext;
-  readmeRel: string;
-  taskId: string;
-  allowPolicy: boolean;
-}): Promise<void> {
-  const stagePaths = new Set<string>([opts.readmeRel]);
-  if (opts.allowPolicy) {
-    const changedPaths = await opts.ctx.git.statusChangedPaths();
-    for (const relPath of changedPaths) {
-      if (
-        protectedPathKindForFile({
-          filePath: relPath,
-          tasksPath: opts.ctx.config.paths.tasks_path,
-          workflowDir: opts.ctx.config.paths.workflow_dir,
-          taskId: opts.taskId,
-        }) === "policy"
-      ) {
-        stagePaths.add(relPath);
-      }
-    }
-  }
-  await opts.ctx.git.stage([...stagePaths].toSorted((a, b) => a.localeCompare(b)));
-}
-
-async function commitRefreshedTaskArtifacts(opts: {
-  ctx: CommandContext;
-  cwd: string;
-  rootOverride?: string;
-  taskId: string;
-  sourceMessage: string;
-  quiet: boolean;
-}): Promise<{ hash: string; subject: string } | null> {
-  const changedPaths = await opts.ctx.git.statusChangedPaths();
-  const taskArtifactPaths = changedPaths.filter((relPath) =>
-    isTaskLocalAdvancePath({
-      workflowDir: opts.ctx.config.paths.workflow_dir,
-      taskId: opts.taskId,
-      tasksPath: opts.ctx.config.paths.tasks_path,
-      relPath,
-    }),
-  );
-  if (taskArtifactPaths.length === 0) return null;
-
-  await stageAllowlist({
-    ctx: opts.ctx,
-    allow: [],
-    allowTasks: true,
-    allowPolicy: false,
-    allowConfig: false,
-    allowHooks: false,
-    allowCI: false,
-    tasksPath: opts.ctx.config.paths.tasks_path,
-    workflowDir: opts.ctx.config.paths.workflow_dir,
-    taskId: opts.taskId,
-    allowTaskOnly: true,
-    emptyAllowMessage:
-      "PR artifact refresh produced no task-local files to stage for the follow-up commit.",
-    noMatchMessage:
-      "PR artifact refresh produced changes outside the active task artifact scope; inspect the working tree before retrying the commit flow.",
-  });
-
-  const message = buildTaskArtifactRefreshCommitSubject({
-    taskId: opts.taskId,
-    baseSubject: opts.sourceMessage,
-  });
-  await guardCommitCheck({
-    ctx: opts.ctx,
-    cwd: opts.cwd,
-    rootOverride: opts.rootOverride,
-    baseBranchOverride: null,
-    taskId: opts.taskId,
-    message,
-    allow: [],
-    allowBase: false,
-    allowTasks: true,
-    allowPolicy: false,
-    allowConfig: false,
-    allowHooks: false,
-    allowCI: false,
-    requireClean: true,
-    quiet: opts.quiet,
-  });
-
-  const env = buildGitCommitEnv({
-    taskId: opts.taskId,
-    allowTasks: true,
-    allowBase: false,
-    allowPolicy: false,
-    allowConfig: false,
-    allowHooks: false,
-    allowCI: false,
-    gitIdentity: await resolveCanonicalGitIdentity(),
-  });
-  await opts.ctx.git.commit({ message, env });
-  return await opts.ctx.git.headHashSubject();
-}
-
-function hasExplicitCommitScope(opts: {
-  allow: string[];
-  allowTasks: boolean;
-  allowPolicy: boolean;
-  allowConfig: boolean;
-  allowHooks: boolean;
-  allowCI: boolean;
-}): boolean {
-  return (
-    opts.allow.some((prefix) => prefix.trim().length > 0) ||
-    opts.allowTasks ||
-    opts.allowPolicy ||
-    opts.allowConfig ||
-    opts.allowHooks ||
-    opts.allowCI
-  );
-}
-
-function formatCommitRef(commit: { hash: string; subject: string } | null): string {
-  if (!commit) return "";
-  return `${commit.hash?.slice(0, 12) ?? ""} ${commit.subject ?? ""}`.trim();
-}
-
-async function stageActiveTaskArtifactsFromAllowTasks(opts: {
-  ctx: CommandContext;
-  taskId: string;
-  allowTasks: boolean;
-}): Promise<string[]> {
-  if (!opts.allowTasks) return [];
-
-  const changedPaths = await opts.ctx.git.statusChangedPaths();
-  const taskArtifactPaths = changedPaths.filter((relPath) =>
-    isTaskLocalAdvancePath({
-      workflowDir: opts.ctx.config.paths.workflow_dir,
-      taskId: opts.taskId,
-      tasksPath: opts.ctx.config.paths.tasks_path,
-      relPath,
-    }),
-  );
-  if (taskArtifactPaths.length === 0) return [];
-
-  const unique = [...new Set(taskArtifactPaths)].toSorted((a, b) => a.localeCompare(b));
-  await opts.ctx.git.stage(unique);
-  return unique;
-}
 
 export async function cmdCommit(opts: {
   ctx?: CommandContext;
@@ -423,16 +253,14 @@ async function cmdCloseCommit(
     return 0;
   }
   if (opts.closeStageTaskArtifacts === true) {
-    if (opts.closeRefreshTaskArtifacts !== false) {
-      await refreshBranchPrArtifactsAfterTaskCommit({
-        ctx: opts.ctx,
-        cwd: opts.cwd,
-        rootOverride: opts.rootOverride,
-        taskId: opts.taskId,
-        quiet: opts.quiet,
-      });
-    }
-    opts.ctx.git.invalidateStatus();
+    await refreshBranchPrArtifactsForCloseCommit({
+      ctx: opts.ctx,
+      cwd: opts.cwd,
+      rootOverride: opts.rootOverride,
+      taskId: opts.taskId,
+      quiet: opts.quiet,
+      refreshTaskArtifacts: opts.closeRefreshTaskArtifacts !== false,
+    });
   }
   await (opts.closeStageTaskArtifacts === true
     ? stageAllowlist({
